ref: Move isSemverRelease & formatVersion into utils/versions/
