+++ static/app/components/group/releaseChart.tsx
@@ -9,7 +9,7 @@ import * as SidebarSection from 'sentry/components/sidebarSection';
 import {t} from 'sentry/locale';
 import type {Group, Release, TimeseriesValue} from 'sentry/types';
 import {getFormattedDate} from 'sentry/utils/dates';
-import {formatVersion} from 'sentry/utils/formatters';
+import {formatVersion} from 'sentry/utils/versions/formatVersion';
 
 /**
  * Stats are provided indexed by statsPeriod strings.
