+++ static/app/utils/versions/formatVersion.tsx
@@ -1,0 +1,16 @@
+import {Release} from '@sentry/release-parser';
+
+export const formatVersion = (rawVersion: string, withPackage = false) => {
+  try {
+    const parsedVersion = new Release(rawVersion);
+    const versionToDisplay = parsedVersion.describe();
+
+    if (versionToDisplay.length) {
+      return `${versionToDisplay}${withPackage && parsedVersion.package ? `, ${parsedVersion.package}` : ''}`;
+    }
+
+    return rawVersion;
+  } catch {
+    return rawVersion;
+  }
+};
