+++ static/app/views/performance/mobile/appStarts/screenSummary/systemApplicationBreakdown.tsx
@@ -10,11 +10,11 @@ import {space} from 'sentry/styles/space';
 import type {TableDataRow} from 'sentry/utils/discover/discoverQuery';
 import EventView from 'sentry/utils/discover/eventView';
 import {DiscoverDatasets} from 'sentry/utils/discover/types';
-import {formatVersion} from 'sentry/utils/formatters';
 import {decodeScalar} from 'sentry/utils/queryString';
 import {MutableSearch} from 'sentry/utils/tokenizeSearch';
 import {useLocation} from 'sentry/utils/useLocation';
 import usePageFilters from 'sentry/utils/usePageFilters';
+import {formatVersion} from 'sentry/utils/versions/formatVersion';
 import {prepareQueryForLandingPage} from 'sentry/views/performance/data';
 import Breakdown from 'sentry/views/performance/mobile/appStarts/screens/breakdown';
 import {useTableQuery} from 'sentry/views/performance/mobile/screenload/screens/screensTable';
