+++ static/app/views/releases/detail/commitsAndFiles/commits.tsx
@@ -15,12 +15,12 @@ import SentryDocumentTitle from 'sentry/components/sentryDocumentTitle';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import type {Commit, Project, Repository} from 'sentry/types';
-import {formatVersion} from 'sentry/utils/formatters';
 import {useApiQuery} from 'sentry/utils/queryClient';
 import routeTitleGen from 'sentry/utils/routeTitle';
 import {useLocation} from 'sentry/utils/useLocation';
 import useOrganization from 'sentry/utils/useOrganization';
 import {useParams} from 'sentry/utils/useParams';
+import {formatVersion} from 'sentry/utils/versions/formatVersion';
 
 import {getCommitsByRepository, getQuery, getReposToRender} from '../utils';
